Refactoring of API calls to improve typing, removed unused methods
