+++ src/lib/server/api/v1/admin_api.js
@@ -5,30 +5,9 @@ import { getLogger } from '$lib/server/logger.js';
 const logger = getLogger('admin API [v1]');
 
 /**
- * Fetches the list of jobs from the server
- * @param {typeof fetch} fetch
- * @returns {Promise<*>}
- */
-export async function getJobs(fetch) {
-	logger.debug('Fetching jobs');
-
-	const response = await fetch(FRACTAL_SERVER_HOST + `/admin/v1/job/?log=false`, {
-		method: 'GET',
-		credentials: 'include'
-	});
-
-	if (response.ok) {
-		return await response.json();
-	}
-
-	logger.error('Unable to fetch the list of jobs');
-	await responseError(response);
-}
-
-/**
  * Fetches the list of projects from the server
  * @param {typeof fetch} fetch
- * @returns {Promise<*>}
+ * @returns {Promise<import('$lib/types').Project[]>}
  */
 export async function listProjects(fetch) {
 	logger.debug('Fetching the list of projects');
@@ -37,10 +16,10 @@ export async function listProjects(fetch) {
 		credentials: 'include'
 	});
 
-	if (response.ok) {
-		return await response.json();
+	if (!response.ok) {
+		logger.error('Unable to fetch the list of projects');
+		await responseError(response);
 	}
 
-	logger.error('Unable to fetch the list of projects');
-	await responseError(response);
+	return await response.json();
 }
