+++ src/lib/server/api/v2/workflow_api.js
@@ -5,23 +5,24 @@ import { getLogger } from '$lib/server/logger.js';
 const logger = getLogger('workflow API [v2]');
 
 /**
- * Fetches all the workflows of a user
+ * Fetches the list of workflows of a project from the server
  * @param {typeof fetch} fetch
- * @returns {Promise<*>}
+ * @param {number|string} projectId
+ * @returns {Promise<Array<import('$lib/types-v2').WorkflowV2>>}
  */
-export async function getUserWorkflows(fetch) {
-	logger.debug('Fetching user workflows');
-	const response = await fetch(FRACTAL_SERVER_HOST + `/api/v2/workflow/`, {
+export async function getWorkflows(fetch, projectId) {
+	logger.debug('Fetching project workflows [project_id=%d]', projectId);
+	const response = await fetch(FRACTAL_SERVER_HOST + `/api/v2/project/${projectId}/workflow/`, {
 		method: 'GET',
 		credentials: 'include'
 	});
 
-	if (response.ok) {
-		return await response.json();
+	if (!response.ok) {
+		logger.error('Unable to fetch project workflows [project_id=%d]', projectId);
+		await responseError(response);
 	}
 
-	logger.debug('Unable to fetch user workflows');
-	await responseError(response);
+	return await response.json();
 }
 
 /**
@@ -29,7 +30,7 @@ export async function getUserWorkflows(fetch) {
  * @param {typeof fetch} fetch
  * @param {string} projectId
  * @param {string} workflowId
- * @returns {Promise<*>}
+ * @returns {Promise<import('$lib/types-v2').WorkflowV2>}
  */
 export async function getWorkflow(fetch, projectId, workflowId) {
 	logger.debug('Fetching workflow [workflow_id=%d] [project_id=%d]', workflowId, projectId);
@@ -41,12 +42,16 @@ export async function getWorkflow(fetch, projectId, workflowId) {
 		}
 	);
 
-	if (response.ok) {
-		return await response.json();
+	if (!response.ok) {
+		logger.error(
+			'Unable to fetch workflow [workflow_id=%d] [project_id=%d]',
+			workflowId,
+			projectId
+		);
+		await responseError(response);
 	}
 
-	logger.error('Unable to fetch workflow [workflow_id=%d] [project_id=%d]', workflowId, projectId);
-	await responseError(response);
+	return await response.json();
 }
 
 /**
@@ -54,7 +59,7 @@ export async function getWorkflow(fetch, projectId, workflowId) {
  * @param {typeof fetch} fetch
  * @param {number|string} projectId
  * @param {number|string} workflowId
- * @returns {Promise<*>}
+ * @returns {Promise<Array<import('$lib/types-v2').ApplyWorkflowV2>>}
  */
 export async function getWorkflowJobs(fetch, projectId, workflowId) {
 	logger.debug('Fetching workflow jobs [workflow_id=%d] [project_id=%d]', workflowId, projectId);
@@ -66,12 +71,12 @@ export async function getWorkflowJobs(fetch, projectId, workflowId) {
 		}
 	);
 
-	if (response.ok) {
-		return await response.json();
+	if (!response.ok) {
+		logger.error(
+			`Unable to fetch workflow jobs [workflow_id=${workflowId}] [project_id=${projectId}]`
+		);
+		await responseError(response);
 	}
 
-	logger.error(
-		`Unable to fetch workflow jobs [workflow_id=${workflowId}] [project_id=${projectId}]`
-	);
-	await responseError(response);
+	return await response.json();
 }
