+++ src/routes/v1/admin/jobs/+page.server.js
@@ -1,11 +1,13 @@
 import { listUsers } from '$lib/server/api/v1/auth_api.js';
 import { listProjects } from '$lib/server/api/v1/admin_api.js';
+import { getLogger } from '$lib/server/logger';
+
+const logger = getLogger('admin jobs page [v1]');
 
 export async function load({ fetch }) {
-	/** @type {import('$lib/types').Project[]} */
-	const projects = await listProjects(fetch);
+	logger.trace('Loading admin jobs page');
 
-	/** @type {import('$lib/types').User[]} */
+	const projects = await listProjects(fetch);
 	const users = await listUsers(fetch);
 
 	return {
