+++ src/routes/v1/jobs/+page.server.js
@@ -5,11 +5,9 @@ import {
 import { getUserJobs, listProjects } from '$lib/server/api/v1/project_api';
 
 export async function load({ fetch }) {
-	/** @type {import('$lib/types').Project[]} */
 	const projects = await listProjects(fetch);
 	sortProjectsByTimestampCreatedDesc(projects);
 
-	/** @type {import('$lib/types').ApplyWorkflow[]} */
 	const jobs = await getUserJobs(fetch);
 
 	const workflows = removeDuplicatedItems(
