+++ src/routes/v1/projects/+page.server.js
@@ -4,12 +4,10 @@ import { getLogger } from '$lib/server/logger.js';
 
 const logger = getLogger('projects page [v1]');
 
-/** @type {import('./$types').PageServerLoad} */
 export async function load({ fetch }) {
 	logger.trace('Loading projects page');
 
 	// Load projects from server
-	/** @type {import('$lib/types').Project[]} */
 	const projects = await listProjects(fetch);
 	sortProjectsByTimestampCreatedDesc(projects);
