+++ src/routes/v1/tasks/+page.server.js
@@ -2,7 +2,6 @@ import { orderTasksByOwnerThenByNameThenByVersion } from '$lib/common/component_
 import { listTasks } from '$lib/server/api/v1/task_api';
 
 export async function load({ fetch }) {
-	/** @type {import('$lib/types').Task[]} */
 	const tasks = await listTasks(fetch);
 
 	orderTasksByOwnerThenByNameThenByVersion(tasks, null, 'desc');
