+++ src/routes/v2/admin/users/[userId]/edit/+page.server.js
@@ -3,11 +3,9 @@ import { getLogger } from '$lib/server/logger.js';
 
 const logger = getLogger('admin edit user page');
 
-/** @type {import('./$types').PageServerLoad} */
 export async function load({ fetch, params }) {
 	logger.trace('Loading user %d', params.userId);
 
-	/** @type {import('$lib/types').User} */
 	const user = await getUser(fetch, params.userId);
 
 	return {
